% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing.R
\name{LOWESS}
\alias{LOWESS}
\title{Locally Weighted Scatterplot Smoothing}
\usage{
LOWESS(x, fps = 25, h = 10, d = 2, r = 2, smooth.para = 6)
}
\arguments{
\item{h}{- Half window size, the window size is 2 * h  + 1}

\item{d}{- Degree of polynomial}

\item{r}{- Number of repeated iterations}

\item{smooth.para}{- Smoothing parameter, defines}

\item{-}{fps, frame per second}
}
\value{
loc - the smoothed location

v - velocity obtained by deriving the polynomial at the specific point

a - accelration, obtained by deriving the polynomial at the specific point twice
}
\description{
Applies LOWESS on select vector as a function of time
}
