% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing.R
\name{RSEE.smooth}
\alias{RSEE.smooth}
\title{RSEE.smooth - performs LOWESS and RRM}
\usage{
RSEE.smooth(object = NULL, path = NULL, project = "",
  individual = "", forced = T, fps = 25, h.loess = 10,
  d.loess = 2, r.loess = 2, frames.rrm = 5, cutoff.rrm = 1e-06,
  h.rrm = c(7, 5, 3, 3))
}
\arguments{
\item{object}{data frame with x and y coordinates.
name of x coordinates vector should contain "x" or "X"
 column name of y coordinates vector should contain "y" or "Y"}

\item{path}{if object is NULL, loads a CSV file of XY coordinates}

\item{project}{project name}

\item{individual}{name of individual animal}

\item{forced}{set to TRUE if no fixed homebase, and FALSE if fixed homebase (free exploration)}

\item{fps}{frames per second (of the tracked video)}

\item{h.loess}{half window width of LOWESS}

\item{d.loess}{degree of polynomial (only d=2 implemented for velocities and accelerations)}

\item{r.loess}{number of LOWESS iterations (recommended: 2)}

\item{frames.rrm}{RRM arguments (see NOTE)}

\item{cutoff.rrm}{RRM arguments (see NOTE)}

\item{h.rrm}{sequence of half-window widths for repeated RM}
}
\value{
The function returns a list with following objects:
 \describe{
 \item{smooth.data}{ data frame, containing for each frame: smooth x and y coordinates,
   smooth velocities and accelerations in x and y, and entry.}
 \item{speed.acc}{data frame of smooth point speeds and accelerations}
 \item{entry}{data frame of start and end frames for each entry (only 1 entry if forced = T)}
 \item{motion.segments}{data frame of start and end frames for each arrest}
 \describe{info - list of general information about the session:
 \item{project, individual}{input names}
 \item{session.length.minutes}{length of the session in minutes}
 \item{home}{fixed homebase coordinates (if forced = F)}
 \item{number.of.entries}{more than 1 if forced = F}
 }
 }
}
\description{
RSEE.smooth - performs LOWESS and RRM
}
\note{
an arrest is a sequence of at least "frames.rrm" frames in which the maximum distance between the two farthest points (after RRM) is no more than "cutoff.rrm"
}
