% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{RSEEsegment}
\alias{RSEEsegment}
\title{RSEEsegment - performs segmentation to lingering and progression}
\usage{
RSEEsegment(object, initial.n.gauss = 2, which.intersect = 0,
  min.length.motion.seg = 2, user.permission = F)
}
\arguments{
\item{object}{list of results from RSEEsmooth}

\item{initial.n.gauss}{number of initial Gaussians to fit (recommended 2)}

\item{which.intersect}{if zero, returns the intersection between the leftmost and the rightmost Gaussians.
In case of two Gaussians, just the intersection between them)
In case of more than two Gaussians, the user can choose which one (from left to right)}

\item{min.length.motion.seg}{minimum length (in frames) of motions segments to use for segmentation}

\item{user.permission}{if TRUE the user gets to verify the segmentation results,}
}
\value{
object - updates input list "object" with the following components:
\describe{
\item{progression.lingering}{an additional data frame including start and end time of each segment,
    prog - 1 if progression, 0 if lingering. entry - in which entry was the segment performed}
\item{motion.segments}{updated to include classification to lingering and progression (prog = 0 or 1)}
\item{info}{updated to include velocity.threshold.cm.s - the velocity threshold between lingering and
    progression (cm/s, non transformed)}}
}
\description{
RSEEsegment - performs segmentation to lingering and progression
}
