% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap.R
\name{SmoothDwell}
\alias{SmoothDwell}
\title{HeatMaker - Calculates smoothed dwell time at each coodinates.}
\usage{
SmoothDwell(smooth.data, x.vec = NULL, y.vec = NULL, sigma = 9,
  n = 57, quant = 0.95, to.plot = FALSE)
}
\arguments{
\item{smooth.data}{A data.frame contains x and y coordinates
name of x coordinates vector should contain "x" or "X"
 column name of y coordinates vector should contain "y" or "Y"}

\item{x.vec}{The grid of x-axis for the heatmap. Defaults to seq(min(x), max(x), 1)}

\item{y.vec}{The grid of y-axis for the heatmap. Defaults to seq(min(y), max(y), 1)}

\item{sigma}{The standard deviation of the Gaussian smoothing kernel, controls the level of the
smoothing (larger, the heatmap will be smoother)}

\item{n}{The filter size is n * n}

\item{quant}{Quantile threshold, returns the > quant local maxima}

\item{to.plot}{Flag, if TRUE will also plot the heatmap}
}
\value{
The function returns a list with following objects:
 \describe{
 \item{smooth.dwell.time}{ a data.frame, x, y, smooth.dwell.time. For each cell coordinates the smoothed dwell time spent}
 \item{maxpoint}{coordinates of local maxima above the predetermined threshold}}
}
\description{
HeatMaker - Calculates smoothed dwell time at each coodinates.
}
